v0.6   - added c-simulator. Incomplete (CELL_SHL/SHR. Nothing with IO yet)
git-svn-id: https://apollo.arh.pub.ro:8443/svn/ConnexSIFT/trunk/software/connex@70 0e9f8c0c-05a5-1044-940b-b810a0af9d24

+++ connexppIAsm/src/io_unit.cpp
@@ -45,7 +45,7 @@ void io_unit::prewriteVectors(UINT16 destAddress, UINT16 *srcAddress, UINT16 num
 
 int io_unit::vwrite()
 {
-    if (Size == write(vpipe_write_32, &Iou,Size))
+    if ((int)Size == write(vpipe_write_32, &Iou,Size))
         return PASS;
     else return FAIL;
 }
@@ -59,7 +59,7 @@ int io_unit::vread()
 {
     if (DESCRIPTOR_SIZE == write(vpipe_write_32, &Iou.Descriptor, DESCRIPTOR_SIZE))
     {
-        if (Size == read(vpipe_read_32, Iou.Content, Size))
+        if ((int)Size == read(vpipe_read_32, Iou.Content, Size))
             return PASS;
     }
